rag/rag_store.py
from langchain_community.document_loaders import TextLoader
from langchain_text_splitters import RecursiveCharacterTextSplitter
from langchain_community.vectorstores import FAISS
from langchain_community.embeddings.fastembed import FastEmbedEmbeddings

def load_rag():
    loader = TextLoader("rag/rag.txt")
    docs = loader.load()

    splitter = RecursiveCharacterTextSplitter(chunk_size=500, chunk_overlap=50)
    splits = splitter.split_documents(docs)

    embeddings = FastEmbedEmbeddings()

    vectordb = FAISS.from_documents(splits, embeddings)
    return vectordb
